--- src/main/java/org/cryptomator/common/mount/ActualMountService.java
@@ -1,6 +1,0 @@
-package org.cryptomator.common.mount;
-
-import org.cryptomator.integrations.mount.MountService;
-
-public record ActualMountService(MountService service, boolean isDesired) {
-}
